Add initial support for QUIC, improve logging capabilities.
Add TRACE logging mode
